chore: set empty hash to hash of nothing (#20775)
