+++ server/v2/cometbft/abci.go
@@ -2,6 +2,7 @@ package cometbft
 
 import (
 	"context"
+	"crypto/sha256"
 	"errors"
 	"fmt"
 	"sync/atomic"
@@ -103,7 +104,7 @@ func (c *Consensus[T]) CheckTx(ctx context.Context, req *abciproto.CheckTxReques
 		return nil, err
 	}
 
-	cometResp := &abci.CheckTxResponse{
+	cometResp := &abciproto.CheckTxResponse{
 		Code:      resp.Code,
 		GasWanted: uint64ToInt64(resp.GasWanted),
 		GasUsed:   uint64ToInt64(resp.GasUsed),
@@ -137,7 +138,7 @@ func (c *Consensus[T]) Info(ctx context.Context, _ *abciproto.InfoRequest) (*abc
 		return nil, err
 	}
 
-	return &abci.InfoResponse{
+	return &abciproto.InfoResponse{
 		Data:    c.cfg.Name,
 		Version: c.cfg.Version,
 		// AppVersion:       cp.GetVersion().App,
@@ -226,11 +227,19 @@ func (c *Consensus[T]) InitChain(ctx context.Context, req *abciproto.InitChainRe
 		})
 	}
 
+	ci, err := c.store.LastCommitID()
+	if err != nil {
+		return nil, err
+	}
+
+	// populate hash with empty byte slice instead of nil
+	bz := sha256.Sum256([]byte{})
+
 	br := &coreappmgr.BlockRequest[T]{
 		Height:            uint64(req.InitialHeight - 1),
 		Time:              req.Time,
-		Hash:              nil,
-		AppHash:           nil,
+		Hash:              bz[:],
+		AppHash:           ci.Hash,
 		ChainId:           req.ChainId,
 		ConsensusMessages: consMessages,
 		IsGenesis:         true,
@@ -271,7 +280,7 @@ func (c *Consensus[T]) InitChain(ctx context.Context, req *abciproto.InitChainRe
 		return nil, fmt.Errorf("unable to write the changeset: %w", err)
 	}
 
-	return &abci.InitChainResponse{
+	return &abciproto.InitChainResponse{
 		ConsensusParams: req.ConsensusParams,
 		Validators:      validatorUpdates,
 		AppHash:         stateRoot,
@@ -317,7 +326,7 @@ func (c *Consensus[T]) PrepareProposal(
 		encodedTxs[i] = tx.Bytes()
 	}
 
-	return &abci.PrepareProposalResponse{
+	return &abciproto.PrepareProposalResponse{
 		Txs: encodedTxs,
 	}, nil
 }
@@ -350,13 +359,13 @@ func (c *Consensus[T]) ProcessProposal(
 	err := c.processProposalHandler(ciCtx, c.app, decodedTxs, req)
 	if err != nil {
 		c.logger.Error("failed to process proposal", "height", req.Height, "time", req.Time, "hash", fmt.Sprintf("%X", req.Hash), "err", err)
-		return &abci.ProcessProposalResponse{
-			Status: abci.PROCESS_PROPOSAL_STATUS_REJECT,
+		return &abciproto.ProcessProposalResponse{
+			Status: abciproto.PROCESS_PROPOSAL_STATUS_REJECT,
 		}, nil
 	}
 
-	return &abci.ProcessProposalResponse{
-		Status: abci.PROCESS_PROPOSAL_STATUS_ACCEPT,
+	return &abciproto.ProcessProposalResponse{
+		Status: abciproto.PROCESS_PROPOSAL_STATUS_ACCEPT,
 	}, nil
 }
 
@@ -529,7 +538,7 @@ func (c *Consensus[T]) VerifyVoteExtension(
 	resp, err := c.verifyVoteExt(ctx, latestStore, req)
 	if err != nil {
 		c.logger.Error("failed to verify vote extension", "height", req.Height, "err", err)
-		return &abci.VerifyVoteExtensionResponse{Status: abci.VERIFY_VOTE_EXTENSION_STATUS_REJECT}, nil
+		return &abciproto.VerifyVoteExtensionResponse{Status: abciproto.VERIFY_VOTE_EXTENSION_STATUS_REJECT}, nil
 	}
 
 	return resp, err
@@ -565,7 +574,7 @@ func (c *Consensus[T]) ExtendVote(ctx context.Context, req *abciproto.ExtendVote
 	resp, err := c.extendVote(ctx, latestStore, req)
 	if err != nil {
 		c.logger.Error("failed to verify vote extension", "height", req.Height, "err", err)
-		return &abci.ExtendVoteResponse{}, nil
+		return &abciproto.ExtendVoteResponse{}, nil
 	}
 
 	return resp, err
